Modularized agent code
